Segregates User Endpoints into services and controllers
